added authetication

+++ src/resolvers/Mutation.js
@@ -1,42 +1,68 @@
-import v4 from 'uuid'
-
-// Enum
-// 1. A special type that defines a set of constants
-// 2. This type can then be used for a field
-// 3. values for the field must be one of the constants for the type
-// UserRole - standard, editor, admin
-// type User {
-// role: UserRole!
-// }
-
+import bcrypt from 'bcryptjs'
+import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 const Mutation = {
   createUser: async (parent, args, { prisma }, info) => {
-    const emailTaken = await prisma.exists.User({ email: args.data.email })
-    if (emailTaken) {
-      throw new Error('E-mail taken')
+    if (args.data.password.length < 8) {
+      throw new Error('Password must be 8 Characters or longer')
     }
-    return prisma.mutation.createUser({ data: args.data }, info)
+
+    // const emailTaken = await prisma.exists.User({ email: args.data.email })
+    // if (emailTaken) {
+    //   throw new Error('E-mail taken')
+    // }
+    const password = await bcrypt.hash(args.data.password, 10)
+    const user = await prisma.mutation.createUser({
+      data: {
+        ...args.data,
+        password
+      }
+    })
+    return {
+      user,
+      token: jwt.sign({ userId: user.id }, 'thisisasecret')
+    }
   },
-  deleteUser: async (parent, args, { prisma }, info) => {
-    const userExists = await prisma.exists.User({ id: args.id })
-    if (!userExists) throw new Error('User not found')
-    return prisma.mutation.deleteUser({ where: { id: args.id } }, info)
+  loginUser: async (
+    parent,
+    { data: { email, password } },
+    { prisma },
+    info
+  ) => {
+    const user = await prisma.query.user({ where: { email: email } })
+    if (!user) throw new Error('User not found')
+    const isMatch = await bcrypt.compare(password, user.password)
+    if (!isMatch) throw new Error('Unable to login')
+
+    return { user, token: jwt.sign({ userId: user.id }, 'thisisasecret') }
   },
-  deletePost: async (parent, args, { prisma }, info) => {
-    return prisma.mutation.deletePost({ where: { id: args.id } }, info)
+  deleteUser: async (parent, args, { prisma, request }, info) => {
+    const userId = getUserId(request)
+    return prisma.mutation.deleteUser({ where: { id: userId } }, info)
   },
-  updateUser: async (parent, args, { prisma }, info) => {
+  updateUser: async (parent, args, { prisma, request }, info) => {
+    const userId = getUserId(request)
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id
+          id: userId
         },
         data: args.data
       },
       info
     )
   },
-  updatePost: async (parent, args, { prisma }, info) => {
+  updatePost: async (parent, args, { prisma, request }, info) => {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+    if (!postExists) {
+      throw new Error('Unable to update post')
+    }
     return prisma.mutation.updatePost(
       {
         data: args.data,
@@ -47,7 +73,30 @@ const Mutation = {
       info
     )
   },
-  updateComment: (parent, args, { prisma }, info) => {
+  deletePost: async (parent, args, { prisma, request }, info) => {
+    const userId = getUserId(request)
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+    if (!postExists) {
+      throw new Error('Unable to delete post')
+    }
+    return prisma.mutation.deletePost({ where: { id: args.id } }, info)
+  },
+  updateComment: async (parent, args, { prisma, request }, info) => {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+    if (!commentExists) {
+      throw Error('Unable to Update comment')
+    }
     return prisma.mutation.updateComment(
       {
         where: { id: args.id },
@@ -56,7 +105,18 @@ const Mutation = {
       info
     )
   },
-  deleteComment: (parent, args, { prisma }, info) => {
+  deleteComment: async (parent, args, { prisma, request }, info) => {
+    const userId = getUserId(request)
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+    // TODO give admin access
+    if (!commentExists) {
+      throw new Error('Could not delete comment')
+    }
     return prisma.mutation.deleteComment(
       {
         where: { id: args.id }
@@ -64,7 +124,8 @@ const Mutation = {
       info
     )
   },
-  createPost: async (parent, args, { prisma }, info) => {
+  createPost: async (parent, args, { prisma, request }, info) => {
+    const userId = getUserId(request)
     return prisma.mutation.createPost(
       {
         data: {
@@ -73,7 +134,7 @@ const Mutation = {
           published: args.data.published,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
           }
         }
@@ -81,14 +142,15 @@ const Mutation = {
       info
     )
   },
-  createComment: async (parent, args, { prisma }, info) => {
+  createComment: async (parent, args, { prisma, request }, info) => {
+    const userId = getUserId(request)
     return prisma.mutation.createComment(
       {
         data: {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
           },
           post: {
